requirements.txt
litellm>=1.55.0
python-dotenv==1.0.0
requests==2.31.0
pymexc
pandas
flask==3.0.0
